@@ -1,28 +1,40 @@
 {
   "name": "dbos",
   "version": "0.0.1",
+  "type": "module",
   "scripts": {
     "build": "tsc",
-    "test": "npx dbos migrate && jest --detectOpenHandles",
-    "lint": "eslint src",
-    "lint-fix": "eslint --fix src"
+    "test": "npx dbos migrate && node --experimental-vm-modules node_modules/jest/bin/jest.js --detectOpenHandles --forceExit",
+    "lint": "eslint .",
+    "lint-fix": "eslint --fix .",
+    "start": "node dist/server.js",
+    "prepare": "husky"
   },
   "devDependencies": {
-    "@dbos-inc/dbos-cloud": "^1.23.8",
-    "@dbos-inc/eslint-plugin": "^3.3.3",
-    "@types/jest": "^29.5.5",
-    "@types/supertest": "^2.0.14",
-    "eslint": "^8.56.0",
+    "@eslint/js": "^9.30.1",
+    "@types/jest": "^30.0.0",
+    "eslint": "^9.30.1",
+    "eslint-config-prettier": "^10.1.5",
+    "globals": "^16.3.0",
+    "husky": "^9.1.7",
     "jest": "^29.7.0",
-    "supertest": "^6.3.3",
+    "lint-staged": "^16.1.2",
+    "prettier": "^3.6.2",
     "ts-jest": "^29.1.1",
-    "typescript": "^5.2.2"
+    "typescript": "^5.2.2",
+    "typescript-eslint": "^8.36.0"
   },
   "dependencies": {
-    "@dbos-inc/dbos-sdk": "^1.28.6",
+    "@dbos-inc/dbos-sdk": "^3.0.27-preview",
+    "@dbos-inc/knex-datasource": "^3.0.27-preview",
+    "@fastify/cors": "^11.0.1",
     "axios": "^1.7.4",
-    "jwks-rsa": "^3.1.0",
-    "koa-jwt": "^4.0.4",
+    "fastify": "^5.4.0",
+    "fastify-jwt-jwks": "^2.0.2",
+    "fastify-raw-body": "^5.0.0",
     "stripe": "^14.22.0"
+  },
+  "lint-staged": {
+    "**/*.{js,jsx,ts,tsx,json,css,scss,md}": "npx prettier --write"
   }
 }
