@@ -1,81 +1,165 @@
-import { createTestingRuntime, TestingRuntime } from "@dbos-inc/dbos-sdk";
-import { CloudSubscription } from "./endpoints";
-import { Utils } from "./subscription";
-import request from "supertest";
+import { FastifyInstance } from 'fastify';
+import { DBOS } from '@dbos-inc/dbos-sdk';
+import knex from 'knex';
+import path from 'path';
 
-describe("cors-tests", () => {
-  let testRuntime: TestingRuntime;
-  const auth0TestID = "testauth0123";
-  const stripeTestID = "teststripe123";
-  const testEmail = "testemail@dbos.dev";
+import { buildEndpoints, findAuth0UserID, findStripeCustomerID, recordStripeCustomer, Utils } from './subscription.js';
+
+describe('subscription-tests', () => {
+  let fastify: FastifyInstance;
 
   beforeAll(async () => {
-    testRuntime = await createTestingRuntime([CloudSubscription, Utils]);
-    await testRuntime.queryUserDB(`DELETE FROM accounts WHERE auth0_subject_id='${auth0TestID}';`);
+    // Set up the database
+    if (!process.env.DBOS_DATABASE_URL) {
+      console.error('DBOS_DATABASE_URL not set!');
+      process.exit(1);
+    }
+    const connectionString = new URL(process.env.DBOS_DATABASE_URL);
+    const appDBName = connectionString.pathname.split('/')[1];
+    connectionString.pathname = '/postgres'; // Set the default database to 'postgres' for initial connection
+    const cwd = process.cwd();
+    const knexConfig = {
+      client: 'pg',
+      connection: connectionString.toString(),
+      migrations: {
+        directory: path.join(cwd, 'migrations'),
+        tableName: 'knex_migrations',
+      },
+    };
+    let knexDB = knex(knexConfig);
+    try {
+      await knexDB.raw(`DROP DATABASE IF EXISTS "${appDBName}" WITH (FORCE);`);
+      await knexDB.raw(`CREATE DATABASE "${appDBName}";`);
+    } finally {
+      await knexDB.destroy();
+    }
+
+    connectionString.pathname = `/${appDBName}`;
+    knexConfig.connection = connectionString.toString();
+    knexDB = knex(knexConfig);
+    try {
+      await knexDB.migrate.latest();
+    } finally {
+      await knexDB.destroy();
+    }
+
+    // Set up the Fastify server and DBOS
+    fastify = await buildEndpoints();
+    DBOS.setConfig({
+      name: 'dbos',
+      databaseUrl: connectionString.toString(),
+    });
+    await DBOS.dropSystemDB();
+    await DBOS.launch();
   });
 
   afterAll(async () => {
-    await testRuntime.destroy();
+    await DBOS.shutdown();
+    await fastify.close();
   });
 
-  test("account-management", async () => {
+  test('account-management', async () => {
     // Check our transactions are correct
-    await expect(testRuntime.invoke(Utils).recordStripeCustomer(auth0TestID, stripeTestID, testEmail)).resolves.toBeFalsy(); // No error
-    await expect(testRuntime.invoke(Utils).findStripeCustomerID(auth0TestID)).resolves.toBe(stripeTestID);
-    await expect(testRuntime.invoke(Utils).findAuth0UserID(stripeTestID)).resolves.toBe(auth0TestID);
-    await expect(testRuntime.invoke(Utils).findAuth0UserID("nonexistent")).rejects.toThrow("Cannot find auth0 user for stripe customer nonexistent"); // Non existent user
+    const auth0TestID = 'testauth0123';
+    const stripeTestID = 'teststripe123';
+    const testEmail = 'testemail@dbos.dev';
+    await expect(recordStripeCustomer(auth0TestID, stripeTestID, testEmail)).resolves.toBeFalsy(); // No error
+    await expect(findStripeCustomerID(auth0TestID)).resolves.toBe(stripeTestID);
+    await expect(findAuth0UserID(stripeTestID)).resolves.toBe(auth0TestID);
+    await expect(findAuth0UserID('nonexistent')).rejects.toThrow(
+      'Cannot find auth0 user for stripe customer nonexistent',
+    ); // Non existent user
   });
 
-  test("subscribe-cors", async () => {
+  test('subscribe-cors', async () => {
     // Check the prefligth request has the correct CORS headers
-    const resp = await request(testRuntime.getHandlersCallback())
-      .options("/subscribe")
-      .set("Origin", "https://dbos.dev")
-      .set("Access-Control-Request-Method", "POST")
-      .set("Authorization", "Bearer testtoken");
-    expect(resp.status).toBe(204);
-    expect(resp.headers["access-control-allow-origin"]).toBe("https://dbos.dev");
-    expect(resp.headers["access-control-allow-credentials"]).toBe("true");
+    fastify.inject(
+      {
+        method: 'OPTIONS',
+        url: '/subscribe',
+        headers: {
+          Origin: 'https://dbos.dev',
+          'Access-Control-Request-Method': 'POST',
+          Authorization: 'Bearer testtoken',
+        },
+      },
+      (err, resp) => {
+        expect(err).toBeFalsy();
+        expect(resp).toBeDefined();
+        expect(resp!.statusCode).toBe(204);
+        expect(resp!.headers['access-control-allow-origin']).toBe('https://dbos.dev');
+        expect(resp!.headers['access-control-allow-credentials']).toBe('true');
+      },
+    );
 
     // Our staging env.
-    const resp2 = await request(testRuntime.getHandlersCallback())
-      .options("/subscribe")
-      .set("Origin", "https://dbos.webflow.io")
-      .set("Access-Control-Request-Method", "POST")
-      .set("Authorization", "Bearer testtoken");
-    expect(resp2.status).toBe(204);
-    expect(resp2.headers["access-control-allow-origin"]).toBe("https://dbos.webflow.io");
-    expect(resp2.headers["access-control-allow-credentials"]).toBe("true");
+    fastify.inject(
+      {
+        method: 'OPTIONS',
+        url: '/subscribe',
+        headers: {
+          Origin: 'https://dbos.webflow.io',
+          'Access-Control-Request-Method': 'POST',
+          Authorization: 'Bearer testtoken',
+        },
+      },
+      (err, resp) => {
+        expect(err).toBeFalsy();
+        expect(resp).toBeDefined();
+        expect(resp!.statusCode).toBe(204);
+        expect(resp!.headers['access-control-allow-origin']).toBe('https://dbos.webflow.io');
+        expect(resp!.headers['access-control-allow-credentials']).toBe('true');
+      },
+    );
 
     // Cloud console
-    const resp3 = await request(testRuntime.getHandlersCallback())
-      .options("/subscribe")
-      .set("Origin", "https://console.dbos.dev")
-      .set("Access-Control-Request-Method", "POST")
-      .set("Authorization", "Bearer testtoken");
-    expect(resp3.status).toBe(204);
-    expect(resp3.headers["access-control-allow-origin"]).toBe("https://console.dbos.dev");
-    expect(resp3.headers["access-control-allow-credentials"]).toBe("true");
+    fastify.inject(
+      {
+        method: 'OPTIONS',
+        url: '/subscribe',
+        headers: {
+          Origin: 'https://console.dbos.dev',
+          'Access-Control-Request-Method': 'POST',
+          Authorization: 'Bearer testtoken',
+        },
+      },
+      (err, resp) => {
+        expect(err).toBeFalsy();
+        expect(resp).toBeDefined();
+        expect(resp!.statusCode).toBe(204);
+        expect(resp!.headers['access-control-allow-origin']).toBe('https://console.dbos.dev');
+        expect(resp!.headers['access-control-allow-credentials']).toBe('true');
+      },
+    );
 
-    const resp4 = await request(testRuntime.getHandlersCallback())
-      .options("/subscribe")
-      .set("Origin", "https://staging.console.dbos.dev")
-      .set("Access-Control-Request-Method", "POST")
-      .set("Authorization", "Bearer testtoken");
-    expect(resp4.status).toBe(204);
-    expect(resp4.headers["access-control-allow-origin"]).toBe("https://staging.console.dbos.dev");
-    expect(resp4.headers["access-control-allow-credentials"]).toBe("true");
+    fastify.inject(
+      {
+        method: 'OPTIONS',
+        url: '/subscribe',
+        headers: {
+          Origin: 'https://staging.console.dbos.dev',
+          'Access-Control-Request-Method': 'POST',
+          Authorization: 'Bearer testtoken',
+        },
+      },
+      (err, resp) => {
+        expect(err).toBeFalsy();
+        expect(resp).toBeDefined();
+        expect(resp!.statusCode).toBe(204);
+        expect(resp!.headers['access-control-allow-origin']).toBe('https://staging.console.dbos.dev');
+        expect(resp!.headers['access-control-allow-credentials']).toBe('true');
+      },
+    );
   });
 
   // Test retrieve cloud credentials
-  test("cloud-credential", async () => {
-    if (!process.env.DBOS_LOGIN_REFRESH_TOKEN || process.env.DBOS_LOGIN_REFRESH_TOKEN == "null") {
-      console.log("Skipping cloud-credentials test, no refresh token provided");
+  test('cloud-credential', async () => {
+    if (!process.env.DBOS_LOGIN_REFRESH_TOKEN || process.env.DBOS_LOGIN_REFRESH_TOKEN == 'null') {
+      console.log('Skipping cloud-credentials test, no refresh token provided');
       return;
     }
     await expect(Utils.retrieveAccessToken()).resolves.toBeTruthy();
-    process.env["DBOS_LOGIN_REFRESH_TOKEN"] = "faketoken";
+    process.env['DBOS_LOGIN_REFRESH_TOKEN'] = 'faketoken';
     await expect(Utils.retrieveAccessToken()).rejects.toThrow();
   });
-
 });
